+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/controller/PartnerServiceController.java
@@ -47,9 +47,9 @@ import io.mosip.pms.partner.manager.constant.PartnerManageEnum;
 import io.mosip.pms.partner.manager.service.PartnerManagerService;
 import io.mosip.pms.partner.request.dto.APIKeyGenerateRequestDto;
 import io.mosip.pms.partner.request.dto.AddContactRequestDto;
+import io.mosip.pms.partner.request.dto.BioExtractorsRequestDto;
 import io.mosip.pms.partner.request.dto.CACertificateRequestDto;
 import io.mosip.pms.partner.request.dto.EmailVerificationRequestDto;
-import io.mosip.pms.partner.request.dto.BioExtractorsRequestDto;
 import io.mosip.pms.partner.request.dto.CredentialTypeRequestDto;
 import io.mosip.pms.partner.request.dto.ExtractorsDto;
 import io.mosip.pms.partner.request.dto.PartnerCertDownloadRequestDto;
@@ -90,9 +90,6 @@ public class PartnerServiceController {
 
 	@Value("${mosip.pms.api.id.partner.exists.post}")
 	private String postPartnerExistsId;
-
-	@Value("${mosip.pms.api.id.partners.bioextractors.request.post:mosip.pms.partners.bioextractors.request.post}")
-	private String postPartnerBioextractorsRequestId;
 
 	@Autowired
 	PartnerService partnerService;
@@ -172,7 +169,7 @@ public class PartnerServiceController {
 	@RequestMapping(value = "/{partnerId}/bioextractors/{policyId}", method = RequestMethod.POST)
 	@Operation(
 			summary = "Service to add bio extractors - deprecated since release-1.3.0-beta.5",
-			description = "This endpoint has been deprecated since the release-1.3.0-beta.5 and replaced by the POST /partners/{partnerId}/policies/{policyId}/bio-extractors-request endpoint.",
+			description = "This endpoint has been deprecated since the release-1.3.0-beta.5 and replaced by the POST /partner-policy-requests/{requestId}/bio-extractors-request endpoint.",
 			deprecated = true)
 	public ResponseEntity<ResponseWrapper<String>> addBiometricExtractors(@PathVariable String partnerId ,@PathVariable String policyId,
 			@RequestBody @Valid RequestWrapper<ExtractorsDto> request){
@@ -182,35 +179,6 @@ public class PartnerServiceController {
 		response.setId(request.getId());
 		response.setVersion(request.getVersion());
 		return new ResponseEntity<>(response, HttpStatus.OK);		
-	}
-
-	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
-	@RequestMapping(value = "/{partnerId}/policies/{policyId}/bio-extractors-request", method = RequestMethod.POST)
-	@Operation(summary = "Service to submit bio extractors request", description = "Persists bio extractor requests against an in-progress partner policy mapping request")
-	public ResponseWrapperV2<String> submitBioExtractorsRequest(
-			@PathVariable String partnerId,
-			@PathVariable String policyId,
-			@RequestBody @Valid RequestWrapperV2<BioExtractorsRequestDto> requestWrapper) {
-		Optional<ResponseWrapperV2<String>> validationResponse =
-				requestValidator.validate(postPartnerBioextractorsRequestId, requestWrapper);
-		if (validationResponse.isPresent()) {
-			return validationResponse.get();
-		}
-		inputValidator.validateRequestInput("partnerId", partnerId);
-		inputValidator.validateRequestInput("policyId", policyId);
-		inputValidator.validateRequestInput("partnerPolicyRequestId", requestWrapper.getRequest().getPartnerPolicyRequestId());
-		requestWrapper.getRequest().getExtractors().forEach(extractor -> {
-			inputValidator.validateRequestInput("attributeName", extractor.getAttributeName());
-			inputValidator.validateRequestInput("biometric", extractor.getBiometric());
-			inputValidator.validateRequestInput("biometricSubTypes", extractor.getBiometricSubTypes());
-			inputValidator.validateRequestInput("extractorProvider", extractor.getExtractorProvider());
-			inputValidator.validateRequestInput("extractorProviderVersion", extractor.getExtractorProviderVersion());
-		});
-		ResponseWrapperV2<String> response = new ResponseWrapperV2<>();
-		response.setResponse(partnerService.submitBioExtractorsRequest(partnerId, policyId, requestWrapper.getRequest()));
-		response.setId(requestWrapper.getId());
-		response.setVersion(requestWrapper.getVersion());
-		return response;
 	}
 
 	@PreAuthorize("hasAnyRole(@authorizedRoles.getPostpartnersbioextractors())")
