+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/manager/service/PartnerManagerService.java
@@ -11,6 +11,7 @@ import io.mosip.pms.partner.manager.dto.*;
 import io.mosip.pms.partner.request.dto.APIKeyGenerateRequestDto;
 import io.mosip.pms.partner.request.dto.APIKeyUpdateRequestDto;
 import io.mosip.pms.partner.request.dto.APIkeyStatusUpdateRequestDto;
+import io.mosip.pms.partner.request.dto.BioExtractorsRequestDto;
 import io.mosip.pms.partner.request.dto.BioextractorConfigurationDeleteRequestDto;
 import io.mosip.pms.partner.request.dto.BioextractorConfigurationRequestDto;
 import io.mosip.pms.partner.request.dto.LinkPolicyGroupRequestDto;
@@ -89,11 +90,13 @@ public interface PartnerManagerService {
 	public String approveRejectPartnerPolicyMapping(String mappingkey, StatusRequestDto statusRequest);
 
 	/**
-	 * Fetch bio-extractor request rows submitted for a partner policy request.
-	 *
-	 * @param requestId business request id (req_id)
-	 * @return bio-extractor request rows
+	 * Submit bio-extractor request rows for a partner policy request.
+	 * `@param` requestId business request id (req_id)
+	 * `@param` request bio-extractor payload
+	 * `@return` submission result identifier/message
 	 */
+	String submitBioExtractorsRequest(String requestId, BioExtractorsRequestDto request);
+
 	BioExtractorsResponseWrapperV2 getPartnerPolicyRequestBioExtractors(String requestId);
 
 	/**
