+++ partner/partner-management-service/src/main/java/io/mosip/pms/partner/service/impl/PartnerServiceImpl.java
@@ -128,7 +128,6 @@ import io.mosip.pms.partner.request.dto.PartnerSearchDto;
 import io.mosip.pms.partner.request.dto.PartnerUpdateDto;
 import io.mosip.pms.partner.request.dto.PartnerUpdateRequest;
 import io.mosip.pms.partner.request.dto.PartnerExistsRequestDto;
-import io.mosip.pms.partner.request.dto.BioExtractorsRequestDto;
 import io.mosip.pms.partner.request.dto.CredentialTypeRequestDto;
 import io.mosip.pms.partner.response.dto.PartnerExistsResponseDto;
 import io.mosip.pms.partner.response.dto.APIkeyRequests;
@@ -547,7 +546,7 @@ public class PartnerServiceImpl implements PartnerService {
 	public RetrievePartnerDetailsResponse getPartnerDetails(String partnerId) {
 		validateLoggedInUserAuthorization(partnerId);
 		RetrievePartnerDetailsResponse response = new RetrievePartnerDetailsResponse();
-		Partner partner = getValidPartner(partnerId, true);
+		Partner partner = partnerHelper.getValidPartner(partnerId, true);
 		response.setPartnerID(partner.getId());
 		// check if the data is encrypted
 		boolean isEncrypted = partner.getEmailIdHash() != null;
@@ -565,23 +564,6 @@ public class PartnerServiceImpl implements PartnerService {
 			response.setPolicyGroup(validateAndGetPolicyGroupById(partner.getPolicyGroupId()).getName());
 		}
 		return response;
-	}
-
-	private Partner getValidPartner(String partnerId, boolean isToRetrieve) {
-		Optional<Partner> partnerById = partnerRepository.findById(partnerId);
-		if (partnerById.isEmpty()) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.RETRIVE_PARTNER_FAILURE, partnerId, "partnerId");
-			throw new PartnerServiceException(ErrorCode.PARTNER_DOES_NOT_EXIST_EXCEPTION.getErrorCode(),
-					ErrorCode.PARTNER_DOES_NOT_EXIST_EXCEPTION.getErrorMessage());
-		}
-		if (!isToRetrieve) {
-			if (!partnerById.get().getIsActive()) {
-				auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.RETRIVE_PARTNER_FAILURE, partnerId, "partnerId");
-				throw new PartnerServiceException(ErrorCode.PARTNER_NOT_ACTIVE_EXCEPTION.getErrorCode(),
-						ErrorCode.PARTNER_NOT_ACTIVE_EXCEPTION.getErrorMessage());
-			}
-		}
-		return partnerById.get();
 	}
 
 	private PolicyGroup validateAndGetPolicyGroupById(String policyGroupId) {
@@ -602,7 +584,7 @@ public class PartnerServiceImpl implements PartnerService {
 			throw new PartnerServiceException(ErrorCode.INVALID_MOBILE_NUMBER_EXCEPTION.getErrorCode(),
 					ErrorCode.INVALID_MOBILE_NUMBER_EXCEPTION.getErrorMessage() + maxMobileNumberLength);
 		};
-		Partner partner = getValidPartner(partnerId, true);
+		Partner partner = partnerHelper.getValidPartner(partnerId, true);
 		if(partnerUpdateRequest.getAdditionalInfo() != null) {
 			isJSONValid(partnerUpdateRequest.getAdditionalInfo().toString());
 		}
@@ -728,7 +710,7 @@ public class PartnerServiceImpl implements PartnerService {
 			contactsFromDb.setUpdDtimes(LocalDateTime.now());
 			resultMessage = "Contacts details updated successfully.";
 		} else {
-			Partner partnerFromDb = getValidPartner(partnerId, false);
+			Partner partnerFromDb = partnerHelper.getValidPartner(partnerId, false);
 			contactsFromDb = new PartnerContact();
 
 			String id = PartnerUtil.createPartnerId();
@@ -805,7 +787,7 @@ public class PartnerServiceImpl implements PartnerService {
 			PartnerCertificateUploadRequestDto partnerCertRequesteDto)
 			throws JsonParseException, JsonMappingException, JsonProcessingException, IOException {
 		validateLoggedInUserAuthorization(partnerCertRequesteDto.getPartnerId());
-		Partner partner = getValidPartner(partnerCertRequesteDto.getPartnerId(), true);
+		Partner partner = partnerHelper.getValidPartner(partnerCertRequesteDto.getPartnerId(), true);
 		if (!partner.getApprovalStatus().equals(PartnerConstants.IN_PROGRESS) && !partner.getIsActive()) {
 			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.RETRIVE_PARTNER_FAILURE, partnerCertRequesteDto.getPartnerId(), "partnerId");
 			throw new PartnerServiceException(ErrorCode.PARTNER_NOT_ACTIVE_EXCEPTION.getErrorCode(),
@@ -1087,121 +1069,6 @@ public class PartnerServiceImpl implements PartnerService {
 	}
 
 	@Override
-	public String submitBioExtractorsRequest(String partnerId, String policyId,
-			BioExtractorsRequestDto extractors) {
-		validateLoggedInUserAuthorization(partnerId);
-		if (extractors == null || extractors.getExtractors() == null || extractors.getExtractors().isEmpty()
-				|| extractors.getPartnerPolicyRequestId() == null || extractors.getPartnerPolicyRequestId().isBlank()) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(ErrorCode.INVALID_PARTNER_INPUT_PARAMETER.getErrorCode(),
-					ErrorCode.INVALID_PARTNER_INPUT_PARAMETER.getErrorMessage());
-		}
-		getValidPartner(partnerId, false);
-
-		String requestId = extractors.getPartnerPolicyRequestId().trim();
-		PartnerPolicyRequest parentPolicyRequest = partnerPolicyRequestRepository.findByPartnerIdAndReqId(partnerId,
-				requestId);
-		boolean invalidParent = parentPolicyRequest == null
-				|| Boolean.TRUE.equals(parentPolicyRequest.getIsDeleted())
-				|| parentPolicyRequest.getPolicyId() == null
-				|| !policyId.equals(parentPolicyRequest.getPolicyId());
-		if (invalidParent) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(ErrorCode.PARTNER_POLICY_REQUEST_NOT_FOUND.getErrorCode(),
-					ErrorCode.PARTNER_POLICY_REQUEST_NOT_FOUND.getErrorMessage());
-		}
-
-		String parentStatus = parentPolicyRequest.getStatusCode();
-		if (!PartnerConstants.IN_PROGRESS.equalsIgnoreCase(parentStatus)) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(
-					ErrorCode.BIOEXTRACT_REQUEST_SEND_PARTNER_POLICY_REQUEST.getErrorCode(),
-					ErrorCode.BIOEXTRACT_REQUEST_SEND_PARTNER_POLICY_REQUEST.getErrorMessage());
-		}
-
-		List<String> createdIds = new ArrayList<>();
-
-		List<String> attributeNames = extractors.getExtractors().stream().map(BioExtractorsDto::getAttributeName).toList();
-		Set<String> uniqueAttributeNames = new HashSet<>();
-		Set<String> uniqueModalities = new HashSet<>();
-		for (BioExtractorsDto extractor : extractors.getExtractors()) {
-			String attrName = extractor.getAttributeName() != null ? extractor.getAttributeName().toLowerCase().trim() : null;
-			String biometric = extractor.getBiometric() != null ? extractor.getBiometric().toLowerCase().trim() : null;
-			
-			if (attrName != null && biometric != null) {
-				if (!uniqueAttributeNames.add(attrName) || !uniqueModalities.add(biometric)) {
-					auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-							"partnerId");
-					throw new PartnerServiceException(ErrorCode.DUPLICATE_EXTRACTOR_CONFIG_IN_REQUEST.getErrorCode(),
-							ErrorCode.DUPLICATE_EXTRACTOR_CONFIG_IN_REQUEST.getErrorMessage());
-				}
-			}
-		}
-
-		if (partnerPolicyBioextractRequestRepository.existsByPartnerPolicyRequestId(parentPolicyRequest.getId())) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(ErrorCode.BIOEXTRACT_REQUEST_ALREADY_EXISTS.getErrorCode(),
-					ErrorCode.BIOEXTRACT_REQUEST_ALREADY_EXISTS.getErrorMessage());
-		}
-		if (extractorProviderRepository.existsByPartnerIdAndPolicyIdAndAttributeNameIn(partnerId, policyId,
-				attributeNames)) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(ErrorCode.DUPLICATE_BIOEXTRACT_DETAIL.getErrorCode(),
-					ErrorCode.DUPLICATE_BIOEXTRACT_DETAIL.getErrorMessage());
-		}
-		for (BioExtractorsDto extractor : extractors.getExtractors()) {
-			validateExtractorForBioExtractRequest(partnerId, extractor);
-			PartnerPolicyBioextractRequest row = new PartnerPolicyBioextractRequest();
-			row.setPartnerPolicyRequestId(parentPolicyRequest.getId());
-			row.setPartId(partnerId);
-			row.setPolicyId(policyId);
-			row.setAttributeName(extractor.getAttributeName());
-			row.setBiometricModality(extractor.getBiometric() == null ? null : extractor.getBiometric().trim());
-			if (extractor.getBiometricSubTypes() != null && !extractor.getBiometricSubTypes().isBlank()) {
-				row.setBiometricSubTypes(extractor.getBiometricSubTypes());
-			}
-			row.setExtractorProvider(extractor.getExtractorProvider());
-			row.setExtractorProviderVersion(extractor.getExtractorProviderVersion() == null ? null : extractor.getExtractorProviderVersion().trim());
-			row.setStatusCode(parentStatus);
-			row.setCrBy(getLoggedInUserId());
-			row.setCrDtimes(Timestamp.valueOf(LocalDateTime.now()));
-			String id = PartnerUtil.generateId();
-			int attempts = 0;
-			while (partnerPolicyBioextractRequestRepository.existsById(id)) {
-				if (attempts >= maxRetries) {
-					LOGGER.error("Failed to generate unique {} (field: '{}') for entity '{}' after {} attempts",
-							"Partner Policy Bioextract Request ID", "id", row.getClass().getSimpleName(), maxRetries);
-					auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-							"partnerId");
-					throw new PartnerServiceException(ErrorCode.UNABLE_TO_GENERATE_UNIQUE_ID.getErrorCode(),
-							String.format(ErrorCode.UNABLE_TO_GENERATE_UNIQUE_ID.getErrorMessage(),
-									"Partner Policy Bioextract Request ID", "id", row.getClass().getSimpleName(),
-									maxRetries));
-				}
-				id = PartnerUtil.generateId();
-				attempts++;
-			}
-			row.setId(id);
-			try {
-				partnerPolicyBioextractRequestRepository.saveAndFlush(row);
-			} catch (DataIntegrityViolationException ex) {
-				auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-						"partnerId");
-				throw new PartnerServiceException(ErrorCode.DUPLICATE_BIOEXTRACT_REQUEST.getErrorCode(),
-						ErrorCode.DUPLICATE_BIOEXTRACT_REQUEST.getErrorMessage());
-			}
-			createdIds.add(id);
-		}
-		auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_SUCCESS, partnerId, "partnerId");
-		return "Bio extract request submitted successfully.";
-	}
-
-	@Override
 	public String submitCredentialTypesRequest(String partnerId, String policyId, CredentialTypeRequestDto request) {
 		validateLoggedInUserAuthorization(partnerId);
 		if (request == null
@@ -1212,7 +1079,7 @@ public class PartnerServiceImpl implements PartnerService {
 					ErrorCode.INVALID_PARTNER_INPUT_PARAMETER.getErrorMessage());
 		}
 
-		getValidPartner(partnerId, false);
+		partnerHelper.getValidPartner(partnerId, false);
 
 		String credentialType = request.getCredentialType().trim();
 		validateCredentialTypes(credentialType);
@@ -1284,52 +1151,6 @@ public class PartnerServiceImpl implements PartnerService {
 
 		auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_CREDENTIAL_TYPE_REQUEST_SUCCESS, partnerId, "partnerId");
 		return "Credential type request submitted successfully.";
-	}
-
-	private void validateExtractorForBioExtractRequest(String partnerId, BioExtractorsDto extractor) {
-		if (extractor == null || extractor.getAttributeName() == null || extractor.getAttributeName().isBlank()
-				|| extractor.getBiometric() == null || extractor.getBiometric().isBlank()
-				|| extractor.getExtractorProvider() == null || extractor.getExtractorProvider().isBlank()
-				|| extractor.getExtractorProviderVersion() == null || extractor.getExtractorProviderVersion().isBlank()) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw new PartnerServiceException(ErrorCode.INVALID_PARTNER_INPUT_PARAMETER.getErrorCode(),
-					ErrorCode.INVALID_PARTNER_INPUT_PARAMETER.getErrorMessage());
-		}
-
-		try {
-			Map<String, String> modalityToAttribute = PartnerUtil.getAllowedBioextractorModalityAttributeNameMap(
-					environment,
-					"mosip.pms.bioextractor.allowed.modalities.attribute.name.map");
-			if (modalityToAttribute == null || modalityToAttribute.isEmpty()) {
-				return;
-			}
-
-			String biometric = extractor.getBiometric().trim().toLowerCase();
-			String attributeName = extractor.getAttributeName().trim().toLowerCase();
-
-			String expectedAttributeName = modalityToAttribute.get(biometric);
-			if (expectedAttributeName == null || expectedAttributeName.isBlank()) {
-				String validModalities = modalityToAttribute.keySet().stream()
-						.reduce((a, b) -> a + ", " + b)
-						.orElse("");
-				throw new PartnerServiceException(
-						ErrorCode.INVALID_INPUT_FORMAT.getErrorCode(),
-						String.format(ErrorCode.INVALID_INPUT_FORMAT.getErrorMessage(), "biometric",
-								"Valid values are: " + validModalities));
-			}
-
-			if (!expectedAttributeName.equalsIgnoreCase(attributeName)) {
-				throw new PartnerServiceException(
-						ErrorCode.INVALID_INPUT_FORMAT.getErrorCode(),
-						String.format(ErrorCode.INVALID_INPUT_FORMAT.getErrorMessage(), "attributeName",
-								"For biometric '" + biometric + "', attributeName must be '" + expectedAttributeName + "'"));
-			}
-		} catch (PartnerServiceException ex) {
-			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.SUBMIT_BIO_EXTRACT_REQUEST_FAILURE, partnerId,
-					"partnerId");
-			throw ex;
-		}
 	}
 
 	/**
@@ -1450,7 +1271,7 @@ public class PartnerServiceImpl implements PartnerService {
 	@Override
 	public String mapPartnerPolicyCredentialType(String credentialType, String partnerId, String policyName) {
 		validateCredentialTypes(credentialType);
-		Partner partner = getValidPartner(partnerId, false);
+		Partner partner = partnerHelper.getValidPartner(partnerId, false);
 		if (!Arrays.stream(credentialTypesRequiredPartnerTypes.split(","))
 				.anyMatch(partner.getPartnerTypeCode()::equalsIgnoreCase)) {
 			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.MAP_POLICY_CREDENTIAL_TYPE_FAILURE, partnerId, "partnerId");
@@ -1969,7 +1790,7 @@ public class PartnerServiceImpl implements PartnerService {
 	 */
 	@Override
 	public String updatePolicyGroup(String partnerId, String policyGroupName) {
-		Partner partner = getValidPartner(partnerId, true);
+		Partner partner = partnerHelper.getValidPartner(partnerId, true);
 		//Approved partners policy group should not be updated
 		if (partner.getIsActive()) {
 			auditUtil.setAuditRequestDto(PartnerServiceAuditEnum.MAP_POLICY_GROUP_FAILURE, partnerId, "partnerId");
@@ -2015,7 +1836,7 @@ public class PartnerServiceImpl implements PartnerService {
 	@Override
 	public PartnerPolicyMappingResponseDto requestForPolicyMapping(PartnerPolicyMappingRequest partnerAPIKeyRequest, String partnerId) {
 		validateLoggedInUserAuthorization(partnerId);
-		Partner partner = getValidPartner(partnerId, false);
+		Partner partner = partnerHelper.getValidPartner(partnerId, false);
 		AuthPolicy authPolicy = validatePolicyGroupAndPolicy(partner.getPolicyGroupId(),
 				partnerAPIKeyRequest.getPolicyName());
 		if(partner.getPolicyGroupId() == null) {
